feat: semver parsing and release commits for python and rust
The rules for semver 2.0.0 have been implemented for python (this was
tested) and rust projects (no idea if this works). A CLI tool was added
that handles intial and regular releases. The rules are currently fixed
to parsing the commit log for a limited set of regexes, but a better
implementation of conventional commits needs implementing.